more fixes for webm
Use maxver, minver and webm to filter out elements that are unsupported
for the current file. This way the file writing doesn't fail because
a deprecated mandatory elemnt from an old version is not present.

Also update the specdata from the latest spec. We really need to
rewrite the specdata parser to parse the new spec format instead.